+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file. This change
 - updated `nextjournal/mardkown` version, which brings Java 22 support
 - add original line-number to generated test vars names (#163) - thanks, @behrica
 - cleaning "\r\n" to avoid a tools.reader bug - thanks, @schneiderlin
-- added support for image urls in `kind/image`
-- added support for video urls in `kind/video`
+- added support for image urls in `:kind/image`
+- added support for video urls in `:kind/video`
+- added support for `:kindly/options` in `:kind/fn`
 
 ## [2-beta16] - 2024-09-14
 - Class or style from kindly/options (#139)